Refactor game and author data
