+++ storage/game_data.rb
@@ -1,6 +1,7 @@
 require_relative '../classes/game'
 require 'json'
 
+def GameData
 def read_games
   data = []
   if File.exist?('data/games.json')
@@ -10,4 +11,19 @@ def read_games
     end
   end
   data
+end
+
+def save_games(games)
+  games_data = []
+  games.each do |game|
+    games_data.push(
+      {
+        multiplayer: game.multiplayer,
+        last_played_at: game.last_played_at,
+        publish_date: game.publish_date
+      }
+    )
+  end
+  File.write('./data/games.json', JSON.generate(games_data))
+end
 end
